SPI byte order changed and comments added.
Changed the byte order in the SPI configuration settings and was thus able to eliminate the need for right-alignment in the array conversion function and reversal of data before conversion. Reversal functions may now be obsolete and there are now two uint8_t -> uint32_t array conversion functions -- one which right-aligns the data, and one which left-aligns the data. The left-align function is more efficient and is now the function being used for data transmission. Win!

Comments were also added. I considered going full Doxygen, but I just don't need that clutter right now. Having well-documented code that's easy for me to read and interpret is far more important at this stage.